Made several changes includix redux and backend

+++ backend/controller/userController.js
@@ -11,9 +11,9 @@ const registerUser = async (req, res) => {
        const user = await User.findOne({email});
 
        if(user){
-              res.status(400).json({message:"User already exists!"});
+              res.status(403).json({message:" Forbidden User already exists!"});
        }else{
-              const newUser = User.create({
+              const newUser = await User.create({
                      name,
                      email,
                      password,
@@ -21,15 +21,14 @@ const registerUser = async (req, res) => {
               });
 
               if (newUser) {
-                     generateToken(res, newUser._id);
-
+                     generateToken(res, newUser._id)
                      res.status(201).json({
-                     _id: newUser._id,
-                     name: newUser.name,
-                     email: newUser.email,
+                            name:newUser.name,
+                            email:newUser.email,
+                            tasks:newUser.tasks
                      });
               } else {
-                     res.status(400);
+                     res.status(500);
                      throw new Error('Invalid user data');
               }
        }
@@ -47,6 +46,7 @@ const loginUser = async (req, res) => {
               res.status(200).json({
                      name:user.name,
                      email:user.email,
+                     tasks:user.tasks
               })
        }else{
               res.status(500).json({message:"Invalid Credentials!"});
@@ -55,14 +55,19 @@ const loginUser = async (req, res) => {
 
 
 const getUser = async (req, res) => {
-       
-       const user = await User.findById(req.user._id);
-
-       if(user){
-              res.status(200).json(user);
-       }else{
-              res.status(401).json({message:"Not Found"});
+       try{
+              console.log(req.user._id);
+              const user = await User.findById(req.user._id);
+              if(user){
+                     res.status(200).json(user);
+              }else{
+                     res.status(404).json({message:"Not Found"});
+              }
+       }catch(e){
+              console.log(e);
        }
+       
+       
 }
 
 const updateUser = async (req,res) => {
@@ -75,9 +80,9 @@ const updateUser = async (req,res) => {
               user.password = password || user.password;
 
               const updatedUser = await user.save();
-              res.status(201).json({name:updatedUser.name,email:updatedUser.email});
+              res.status(201).json({name:updatedUser.name,email:updatedUser.email,tasks:updatedUser.tasks});
        }else{
-              res.status(401).json({message:'User not found!'});
+              res.status(404).json({message:'User not found!'});
        }
 
 }
@@ -91,19 +96,25 @@ const logoutUser = async (req, res) => {
            res.status(200).json({message:'Logged out successfully'})
 
        } catch (error) {
-              res.status(501).json({message:error});
+              res.status(500).json({message:error});
        }   
 }
 
 const deleteUser = async (req, res) => {
 
+       const password = req.query.password;
+
        const user = await User.findById(req.user._id);
 
-       if(user){
+       if(user && (await user.matchPassword(password))){
+              res.cookie("jwt","", {
+                     httpOnly:true,
+                     expires: new Date(0)
+                  })   
               const deletedUser = await User.deleteOne({_id:user._id})
               res.status(200).json({message:'User deleted',user:deletedUser});
        }else{
-              res.status(401).json({message:'User not found!'});
+              res.status(404).json({message:'User not found!'});
        }
 }
 
@@ -112,7 +123,7 @@ const createTask = async (req,res) => {
               const { task_name, start_date, end_date, status } = req.body;
               const user = await User.findById(req.user._id);
               if(user){
-                     user.tasks.push({
+                     user.tasks.unshift({
                             task_name,
                             start_date,
                             end_date,
@@ -120,15 +131,10 @@ const createTask = async (req,res) => {
                      });
 
                      const updatedUser = await user.save();
-                     res.status(201).json({message:"Task Created", task:{
-                            task_name,
-                            start_date,
-                            end_date,
-                            status
-                     } });
+                     res.status(201).json({name:updatedUser.name,email:updatedUser.email,tasks:updatedUser.tasks});
               }
        } catch (error) {
-              res.status(501).json({message:error});
+              res.status(500).json({message:error});
        }
 }
 
@@ -150,13 +156,13 @@ const updateTask = async (req, res) => {
                      const updatedTask = await user.save();
 
                      if(updatedTask){
-                            res.status(201).json({message:"Task Updated Successfully !"});
+                            res.status(200).json({name:user.name,email:user.email,tasks:user.tasks});
                      }else{
-                            res.status(501).json({message:"Server Error!"});
+                            res.status(500).json({message:"Server Error!"});
                      }
               }
        }catch (e){
-              res.status(501).json({message:e});
+              res.status(500).json({message:e});
        }
 }
 
@@ -168,9 +174,9 @@ const deleteTask = async (req, res) => {
               var foundIndex = tasks.findIndex(x => x._id == task_id);
               tasks.splice(foundIndex, 1);
               const upadtedUser = await user.save();
-              res.status(201).json({message:"Task Deleted Successfully !"});
+              res.status(200).json({name:upadtedUser.name,email:upadtedUser.email,tasks:upadtedUser.tasks});
        } catch (error) {
-              res.status(501).json({message:error});
+              res.status(500).json({message:error});
        }
 }
